chore: Added distributed tracing propagation to SNS
