+++ lib/subscribers/aws-sdk/middleware/sns/index.js
@@ -7,6 +7,7 @@
 
 const MessageBrokerDesc = require('#agentlib/message-broker-description.js')
 const genericRecorder = require('#agentlib/metrics/recorders/generic.js')
+const attachHeaders = require('#agentlib/subscribers/aws-sdk/utils/attach-headers.js')
 
 module.exports = {
   init() {
@@ -63,10 +64,12 @@ function snsMiddleware(subscriber, config, next, context) {
       recorder: genericRecorder,
       ctx
     })
-    // In a typical producer type instrumentation we'd invoke
-    // `subscriber.insertDTHeaders` here. But since SNS is not really a message
-    // storage destination, headers cannot be propagated. Therefore, we do not
-    // need to attach them here.
+    const headers = Object.create(null)
+    subscriber.insertDTHeaders({ headers, ctx: newCtx, useMqNames: true })
+    if (Object.hasOwn(command.input, 'MessageAttributes') === false) {
+      command.input.MessageAttributes = {}
+    }
+    attachHeaders({ message: command.input, headers })
 
     const fn = agent.tracer.bindFunction(next, newCtx, true)
     return fn(...args)
