+++ lib/subscribers/aws-sdk/middleware/sqs/index.js
@@ -8,13 +8,14 @@
 const path = require('node:path')
 const MessageBrokerDesc = require('#agentlib/message-broker-description.js')
 const genericRecorder = require('#agentlib/metrics/recorders/generic.js')
+const retrieveHeaders = require('#agentlib/subscribers/aws-sdk/utils/retrieve-headers.js')
+const attachHeaders = require('#agentlib/subscribers/aws-sdk/utils/attach-headers.js')
 
 const COMMANDS_TO_INSTRUMENT = [
   'ReceiveMessageCommand',
   'SendMessageCommand',
   'SendMessageBatchCommand'
 ]
-const DT_HEADERS = ['traceparent', 'tracestate', 'newrelic']
 
 module.exports = {
   init() {
@@ -26,10 +27,6 @@ module.exports = {
     step: 'initialize',
     priority: 'high',
     override: true
-  },
-  internal: {
-    attachHeaders,
-    retrieveHeaders
   }
 }
 
@@ -118,53 +115,4 @@ function urlComponents(queueUrl) {
     return matches.groups
   }
   return { region: undefined, accountId: undefined, queue: undefined }
-}
-
-/**
- * Mutates the `MessageAttributes` attached to `message` by attaching any
- * values attached to the `headers` object as new attributes.
- *
- * @param {object} params Function parameters.
- * @param {object} params.message AWS SQS message object that has
- * `MessageAttributes`.
- * @param {object} params.headers A hash of distributed trace headers to
- * propagate.
- */
-function attachHeaders({ message, headers }) {
-  // SQS allows a maximum of 10 message attributes.
-  const MAX_HEADERS = 10
-  const inputAttrsCount = Object.keys(message.MessageAttributes).length
-
-  // Add headers in priority order. If there isn't enough room in the SQS
-  // message, this ensures we get the most important header(s) sent if possible.
-  const availSlots = MAX_HEADERS - inputAttrsCount
-  let i = 1
-  for (const header of DT_HEADERS) {
-    if (i > availSlots) break
-    if (Object.hasOwn(headers, header) === false) continue
-    message.MessageAttributes[header] = {
-      DataType: 'String',
-      StringValue: headers[header]
-    }
-    i += 1
-  }
-}
-
-/**
- * Finds any distributed trace headers present on the given SQS message
- * and returns them as a headers object.
- *
- * @param {object} params Function params.
- * @param {object} params.message An SQS message instance.
- *
- * @returns {object} Hash of key value pairs.
- */
-function retrieveHeaders({ message }) {
-  const headers = Object.create(null)
-  const attrs = message.MessageAttributes || {}
-  for (const header of DT_HEADERS) {
-    if (Object.hasOwn(attrs, header) === false) continue
-    headers[header] = attrs[header].StringValue
-  }
-  return headers
 }
