+++ lib/subscribers/aws-sdk/utils/constants.js
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2026 New Relic Corporation. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+'use strict'
+
+const DT_HEADERS = ['traceparent', 'tracestate', 'newrelic']
+
+module.exports = {
+  DT_HEADERS
+}
